Moved SwiftMarshal into Swift.Runtime.InteropServices
